requirements.txt
aiohttp==3.11.18
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.3
fastapi==0.115.12
fastapi-utils==0.8.0
Flask==3.1.1
greenlet==3.2.3
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
kafka-python==2.0.5
MarkupSafe==3.0.2
maxminddb==2.7.0
mysql-connector-python==9.2.0
numpy==1.26.4
packaging==25.0
pandas==2.2.3
pika==1.3.2
pyudev==0.24.3
psutil==5.9.8
psycopg2-binary==2.9.10
pydantic==2.11.2
pydantic_core==2.33.1
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
scikit-learn==1.3.2
SQLAlchemy==2.0.41
starlette==0.46.1
threadpoolctl==3.5.0
typing_extensions==4.13.1
tzdata==2025.1
urllib3==2.4.0
uvicorn==0.34.0
watchdog==6.0.0
websocket-client==1.8.0
Werkzeug==3.1.3
tensorflow-cpu==2.20.0 
keras==3.10.0
scipy==1.15.2
